+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 ## Unreleased
 
+- Fix module reloading with different versions. [#254](https://github.com/BernieWhite/PSRule/issues/254)
+- Fix not finding rules in current path by default. [#256](https://github.com/BernieWhite/PSRule/issues/256)
+
 ## v0.8.0-B190742 (pre-release)
 
 - Fix inconsistent handling of `$PWD`. [#249](https://github.com/BernieWhite/PSRule/issues/249)